+++ kern/include/corerequest.h
@@ -25,6 +25,10 @@ struct core_request_data {
 	struct sched_pcore_tailq  prov_not_alloc_me;  /* maybe alloc to others */
 };
 
+/* Provision a core to proc p. This code assumes that the scheduler that uses
+ * it holds a lock for the duration of the call. */
+void __provision_core(struct proc *p, struct sched_pcore *spc);
+
 static inline uint32_t spc2pcoreid(struct sched_pcore *spc)
 {
 	extern struct sched_pcore *all_pcores;
